refactor: 패키지 화면 별로 변경
