+++ README.md
@@ -23,3 +23,6 @@
 - [x] Cinema를 class로 변경
 - [x] onCreate 함수 분리
 - [x] 데이터 바인딩에 필요한 값들만 대입하도록 수정
+
+### 2차 피드백 반영
+- [x] Domain에서 사용하지 않는 프로퍼티 제거
