+++ app/src/main/java/woowacourse/movie/ui/view/MainPresenter.kt
@@ -1,7 +1,4 @@
-package woowacourse.movie.ui.presenter
-
-import woowacourse.movie.ui.contract.MainContract
-import woowacourse.movie.ui.view.MainScreen
+package woowacourse.movie.ui.view
 
 class MainPresenter(private val view: MainContract.View) : MainContract.Presenter {
     override fun presentScreen(mainScreen: MainScreen) {
