+++ app/src/main/java/woowacourse/movie/ui/view/history/ReservationHistoryFragment.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.view.reservation
+package woowacourse.movie.ui.view.history
 
 import android.os.Bundle
 import android.view.LayoutInflater
