+++ app/src/main/java/woowacourse/movie/ui/view/screening/adapter/ScreeningAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.view.cinema.adapter
+package woowacourse.movie.ui.view.screening.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
