+++ app/src/main/res/layout/fragment_reservation_history.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.view.reservation.ReservationHistoryFragment">
+    tools:context=".ui.view.history.ReservationHistoryFragment">
 
     <TextView
         android:layout_width="match_parent"
